notebook/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-debug-toolbar==3.4.0
django-simple-captcha==0.5.17
sqlparse==0.4.2
